pyproject.toml
[project]
name = "pasteir"
version = "0.1.0"
description = "A modern, secure, and privacy-focused pastebin service, built with Django."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "cryptography>=45.0.5",
    "dj-rest-auth>=7.0.1",
    "django>=5.2.4",
    "django-health-check>=3.20.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.0",
    "drf-spectacular>=0.28.0",
    "django-tasks-scheduler>=4.0.5",
    "gunicorn>=23.0.0",
    "psycopg[binary,pool]>=3.2.9 ; sys_platform == 'win32'",
    "psycopg[c,pool]>=3.2.9 ; sys_platform != 'win32'",
    "python-dotenv>=1.1.1",
    "redis>=6.2.0",
    "g4f",
]
